Update 1.5
Had major major issues across all websites with the navigation bar, as well as floats, content overflow, and more.

Took the time to re-organize the code, cleaned it up, and now have new tasks to work on.

Still not clear on the padding elements but I will work on those another time while I try to edit my pages.