Update 3.0
Finally figured out how to get my resume page to look the way I wanted it to.

Still not very clear on how the floats work but I made it work with this project. I had a border around each element, but it doesnt create the sectional look I want in different screen sizes. I think that Is a natural behavior of CSS, so for now I will leave it this way.